Combine LoginWebAPI.java into LoginController.java, merge test cases
