+++ pom.xml
@@ -119,11 +119,11 @@
       <groupId>org.springframework.boot</groupId>
       <artifactId>spring-boot-starter-data-jpa</artifactId>
     </dependency>
-    <!--		<dependency>-->
-    <!--			<groupId>org.projectlombok</groupId>-->
-    <!--			<artifactId>lombok</artifactId>-->
-    <!--			<scope>provided</scope>-->
-    <!--		</dependency>-->
+    <dependency>
+        <groupId>org.projectlombok</groupId>
+        <artifactId>lombok</artifactId>
+        <scope>provided</scope>
+    </dependency>
     <!--		<dependency>-->
     <!--			<groupId>org.springframework.boot</groupId>-->
     <!--			<artifactId>spring-boot-starter-test</artifactId>-->
