+++ src/test/java/login_system/LoginControllerTest.java
@@ -10,6 +10,7 @@ import org.springframework.context.annotation.FilterType;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit4.SpringRunner;
 import teamchalkling.chalkling.login_system.LoginController;
+import teamchalkling.chalkling.login_system.UserJSON;
 import teamchalkling.chalkling.user_system.*;
 
 import static org.junit.Assert.*;
@@ -22,6 +23,12 @@ public class LoginControllerTest {
     private LoginController loginController;
     @Autowired
     private UserRepository userRepository;
+
+    // WebAPI stuff
+    private final UserJSON user0 = new UserJSON("test0", "password123");
+    private final UserJSON user0_sim = new UserJSON("test0", "password1234");
+    private final UserJSON user1 = new UserJSON("test1", "?#$%:3;");
+    private final UserJSON user2 = new UserJSON("test2", "helloworld");
 
     @Before
     public void setUp(){
@@ -37,6 +44,23 @@ public class LoginControllerTest {
         assertTrue(loginController.check("user2", "pass2"));
         assertFalse(loginController.check("user1", "pass2")); // wrong password
         assertFalse(loginController.check("user3", "pass3")); // invalid username
+    }
+
+    @Test
+    public void testSignUp(){
+        assertTrue(loginController.addUser(user0).getStatus());
+        assertFalse(loginController.addUser(user0_sim).getStatus()); // preserve uniqueness
+    }
+
+    @Test
+    public void testLogIn(){
+        assertTrue(loginController.addUser(user0).getStatus());
+        assertTrue(loginController.addUser(user1).getStatus());
+        assertTrue(loginController.verifyLogin(user0).getStatus());
+        assertTrue(loginController.verifyLogin(user1).getStatus());
+
+        assertFalse(loginController.verifyLogin(user2).getStatus());
+        assertFalse(loginController.verifyLogin(user0_sim).getStatus());
     }
 
 }
